Refactor to separate logic out from command layer into a SqlSchemaManager layer.
