+++ tools/SchemaManager.Core/Model/MutuallyExclusiveType.cs
@@ -1,0 +1,16 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+namespace SchemaManager.Core.Model
+{
+    public class MutuallyExclusiveType
+    {
+        public int Version { get; set; }
+
+        public bool Next { get; set; }
+
+        public bool Latest { get; set; }
+    }
+}
