+++ tools/SchemaManager/Commands/CurrentCommand.cs
@@ -4,35 +4,24 @@
 // -------------------------------------------------------------------------------------------------
 
 using System;
-using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.CommandLine.Rendering;
 using System.CommandLine.Rendering.Views;
-using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using EnsureThat;
-using Microsoft.Extensions.Logging;
-using Microsoft.Health.SqlServer.Configs;
-using Microsoft.Health.SqlServer.Features.Schema.Manager;
-using Microsoft.Health.SqlServer.Features.Schema.Manager.Exceptions;
 using Microsoft.Health.SqlServer.Features.Schema.Manager.Model;
+using SchemaManager.Core;
 
 namespace SchemaManager.Commands
 {
     public class CurrentCommand : Command
     {
-        private readonly BaseSchemaRunner _baseSchemaRunner;
-        private readonly SqlServerDataStoreConfiguration _sqlServerDataStore;
-        private readonly ISchemaClient _schemaClient;
-        private readonly ILogger<CurrentCommand> _logger;
+        private readonly ISchemaManager _schemaManager;
 
-        public CurrentCommand(
-            BaseSchemaRunner baseSchemaRunner,
-            SqlServerDataStoreConfiguration sqlServerDataStore,
-            ISchemaClient schemaClient,
-            ILogger<CurrentCommand> logger)
+        public CurrentCommand(ISchemaManager schemaManager)
             : base(CommandNames.Current, Resources.CurrentCommandDescription)
         {
             AddOption(CommandOptions.ServerOption());
@@ -40,23 +29,17 @@ namespace SchemaManager.Commands
 
             Handler = CommandHandler.Create(
                 (InvocationContext context, Uri server, string connectionString, CancellationToken token)
-                => HandlerAsync(context, server, connectionString, token));
+                => HandlerAsync(context, connectionString, server, token));
 
             Argument.AddValidator(symbol => Validators.RequiredOptionValidator.Validate(symbol, CommandOptions.ConnectionStringOption(), Resources.ConnectionStringRequiredValidation));
             Argument.AddValidator(symbol => Validators.RequiredOptionValidator.Validate(symbol, CommandOptions.ServerOption(), Resources.ServerRequiredValidation));
 
-            EnsureArg.IsNotNull(baseSchemaRunner);
-            EnsureArg.IsNotNull(sqlServerDataStore);
-            EnsureArg.IsNotNull(schemaClient);
-            EnsureArg.IsNotNull(logger, nameof(logger));
+            EnsureArg.IsNotNull(schemaManager, nameof(schemaManager));
 
-            _baseSchemaRunner = baseSchemaRunner;
-            _sqlServerDataStore = sqlServerDataStore;
-            _schemaClient = schemaClient;
-            _logger = logger;
+            _schemaManager = schemaManager;
         }
 
-        private async Task HandlerAsync(InvocationContext invocationContext, Uri server, string connectionString, CancellationToken cancellationToken = default)
+        private async Task HandlerAsync(InvocationContext invocationContext, string connectionString, Uri server, CancellationToken cancellationToken = default)
         {
             var region = new Region(
                           0,
@@ -64,38 +47,12 @@ namespace SchemaManager.Commands
                           Console.WindowWidth,
                           Console.WindowHeight,
                           true);
-            List<CurrentVersion> currentVersions = null;
 
-            _schemaClient.SetUri(server);
-
-            try
-            {
-                _sqlServerDataStore.ConnectionString = connectionString;
-
-                // Base schema is required to run the schema migration tool.
-                // This method also initializes the database if not initialized yet.
-                await _baseSchemaRunner.EnsureBaseSchemaExistsAsync(cancellationToken);
-
-                // If InstanceSchema table is just created(as part of baseSchema), it takes a while to insert a version record
-                // since the Schema job polls and upserts at the specified interval in the service.
-                await _baseSchemaRunner.EnsureInstanceSchemaRecordExistsAsync(cancellationToken);
-
-                currentVersions = await _schemaClient.GetCurrentVersionInformationAsync(cancellationToken);
-            }
-            catch (SchemaManagerException ex)
-            {
-                _logger.LogError(ex, ex.Message);
-                return;
-            }
-            catch (HttpRequestException ex)
-            {
-                _logger.LogError(ex, string.Format(Resources.RequestFailedMessage, server));
-                return;
-            }
+            var currentVersions = await _schemaManager.GetCurrentSchema(connectionString, server, cancellationToken);
 
             var tableView = new TableView<CurrentVersion>
             {
-                Items = currentVersions,
+                Items = new ReadOnlyCollection<CurrentVersion>(currentVersions),
             };
 
             tableView.AddColumn(
