+++ tools/SchemaManager/Program.cs
@@ -12,6 +12,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Health.SqlServer;
 using Microsoft.Health.SqlServer.Configs;
 using Microsoft.Health.SqlServer.Features.Schema.Manager;
+using SchemaManager.Core;
 
 namespace SchemaManager
 {
@@ -49,6 +50,7 @@ namespace SchemaManager
             services.AddSingleton<BaseSchemaRunner>();
             services.AddSingleton<ISchemaManagerDataStore, SchemaManagerDataStore>();
             services.AddSingleton<ISchemaClient, SchemaClient>();
+            services.AddSingleton<ISchemaManager, SqlSchemaManager>();
             services.AddLogging(configure => configure.AddConsole());
             return services.BuildServiceProvider();
         }
